python(refactor): rules module
