+++ python/lib/sift_py/ingestion/rule/__init__.py
@@ -1,3 +1,4 @@
 """
-Concerned with loading in rules to use during ingestion.
+This module is deprecated and exists only to re-export members
+of `sift_py.rule`. Prefer to use that module.
 """
