+++ python/lib/sift_py/rule/_config_test.py
